requirements/requirements.txt
Django==5.0
django-shortcuts==1.6
djangorestframework==3.14.0
environs==10.0.0
psycopg2-binary==2.9.9